Refactoring of data loaders
- move from callables loaders to iterables
- make the loader more modular
- reduce duplicated code
