+++ fortuna/data/loader/__init__.py
@@ -1,0 +1,2 @@
+from fortuna.data.loader.base import BaseInputsLoader, BaseTargetsLoader, DeviceDimensionAugmentedLoader
+from fortuna.data.loader.array_loaders import DataLoader, InputsLoader, TargetsLoader
