+++ fortuna/typing.py
@@ -14,9 +14,10 @@ OptaxOptimizer = GradientTransformation
 Array = Union[jnp.ndarray, np.ndarray]
 Status = Dict[str, Array]
 Path = Union[str, pathlib.Path]
-Batch = Tuple[Array, Array]
-Outputs = jnp.ndarray
+InputData = Array
 Targets = Array
+Batch = Tuple[InputData, Targets]
+Outputs = jnp.ndarray
 Uncertainties = jnp.ndarray
 Predictions = jnp.ndarray
 AnyKey = Union[str, int]
